* Correct invocation of ranlib; the ugly truth, from `man ranlib':
  BUGS
       The way libraries use to be created, errors were  possible
       if  the  library  was modified with ar(1) and the table of
       contents was not updated by rerunning ranlib(1).  Thus the
       link  editor,  ld,  warns  when the modification date of a
       library is more recent than the creation date of its table
       of  contents.   Unfortunately, this means that you get the
       warning even if you only copy the library.

* As a result, new-meta now runs on BSD 4.x Unices (including Mac OS X
  Server, the *BSD family, ...)

+++ meta/in-output/in-output.c
@@ -74,13 +74,15 @@ char *slurpfile(char *fnam, size_t *size)
                 return NULL;
         }
         if((buf = (char *)malloc(*size + 1)) == NULL ) {
-                fprintf(stderr, "could not allocate %i bytes for %s\n", *size + 1, fnam);
+                fprintf(stderr, "could not allocate %i bytes for %s\n",
+                        (int) *size+1, fnam);
                 fclose(fd);
                 *size = 0;
                 return NULL;
         }
         if(fread(buf, 1, *size, fd) != *size) {
-                fprintf(stderr, "could not fread() %i bytes %s\n", *size, fnam);
+                fprintf(stderr, "could not fread() %i bytes %s\n",
+                        (int) *size, fnam);
                 free(buf);
                 *size = 0;
                 buf = NULL;
